refactor(models): rename `model.ts` to `base.ts`, update usages
