+++ src/models/PhoneStatus.ts
@@ -1,4 +1,4 @@
-import CCUModel from './model';
+import CCUModel from '@/models/base';
 
 export default class PhoneStatus extends CCUModel<PhoneStatus> {
   static entity = 'phone_statuses';
