+++ src/models/User.ts
@@ -7,7 +7,7 @@ import type { Config } from '@vuex-orm/plugin-axios';
 import { AuthService } from '../services/auth.service';
 import Language from './Language';
 import Role from './Role';
-import CCUModel from '@/models/model';
+import CCUModel from '@/models/base';
 
 export default class User extends CCUModel<User> {
   static entity = 'users';
