+++ src/models/WorksiteRequest.ts
@@ -1,6 +1,6 @@
 import moment from 'moment';
 import type { Config, Request } from '@vuex-orm/plugin-axios';
-import CCUModel from '@/models/model';
+import CCUModel from '@/models/base';
 import type Organization from '@/models/Organization';
 
 export default class WorksiteRequest extends CCUModel<WorksiteRequest> {
